refactor: language classic game
